use energy decomposition for nonbonded validation
