+++ cg_openmm/tests/test_cgmodel.py
@@ -8,15 +8,19 @@ import sys
 
 # Import package, test suite, and other packages as needed
 import cg_openmm
+import mdtraj as md
 import openmm
 import pytest
 from cg_openmm.cg_model.cgmodel import CGModel
+from cg_openmm.simulation.tools import minimize_structure
+from numpy.testing import assert_almost_equal
 from openmm import unit
 from openmm.app.pdbfile import PDBFile
 
 current_path = os.path.dirname(os.path.abspath(__file__))
-data_path = os.path.join(current_path, 'test_structures')
-   
+data_path = os.path.join(current_path, 'test_data')
+structures_path = os.path.join(current_path,'test_structures')
+
 @pytest.fixture
 def create_cgmodel():
     # Coarse grained model settings
@@ -94,7 +98,7 @@ def create_cgmodel():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -433,7 +437,7 @@ def test_exclusions_1_1_1():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -553,7 +557,7 @@ def test_exclusions_0_1_1():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -673,7 +677,7 @@ def test_exclusions_1_0_1():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -795,7 +799,7 @@ def test_exclusions_1_1_0():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -917,7 +921,7 @@ def test_exclusions_0_0_1():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -1040,7 +1044,7 @@ def test_exclusions_0_1_0():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -1163,7 +1167,7 @@ def test_exclusions_1_0_0():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -1288,7 +1292,7 @@ def test_exclusions_0_0_0():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -1415,7 +1419,7 @@ def test_exclusions_0_0_0_rosetta():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -1545,7 +1549,7 @@ def test_exclusions_0_1_1_sc():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -1660,7 +1664,7 @@ def test_sums_periodic_torsions_1():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -1762,7 +1766,7 @@ def test_sums_periodic_torsions_2():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -1865,7 +1869,7 @@ def test_sums_periodic_torsions_3():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -1968,7 +1972,7 @@ def test_sums_periodic_torsions_4():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -2070,7 +2074,7 @@ def test_sums_periodic_torsions_5():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -2177,7 +2181,7 @@ def test_restricted_angle_potential():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -2281,7 +2285,7 @@ def test_cosine_angle_potential():
     
     sequence = 24 * [A]
     
-    pdb_path = os.path.join(data_path, "24mer_1b1s_initial_structure.pdb")
+    pdb_path = os.path.join(structures_path, "24mer_1b1s_initial_structure.pdb")
     positions = PDBFile(pdb_path).getPositions()
     
     # Build a coarse grained model
@@ -2304,3 +2308,252 @@ def test_cosine_angle_potential():
         sequence=sequence,
         monomer_types=[A],
     )
+
+
+def test_set_binary_interaction_trivial(tmpdir):
+    """Regression test for adding binary interaction parameters"""
+
+    # Load in cgmodel
+    cgmodel = pickle.load(open(f"{data_path}/stored_cgmodel.pkl", "rb" ))    
+
+    binary_interaction_parameters = {
+        "bb_sc_binary_interaction": 0.0}
+    
+    cgmodel_new = CGModel(
+        particle_type_list=cgmodel.particle_type_list,
+        bond_lengths=cgmodel.bond_lengths,
+        bond_force_constants=cgmodel.bond_force_constants,
+        bond_angle_force_constants=cgmodel.bond_angle_force_constants,
+        torsion_force_constants=cgmodel.torsion_force_constants,
+        equil_bond_angles=cgmodel.equil_bond_angles,
+        torsion_periodicities=cgmodel.torsion_periodicities,
+        torsion_phase_angles=cgmodel.torsion_phase_angles,
+        binary_interaction_parameters=binary_interaction_parameters,
+        include_nonbonded_forces=cgmodel.include_nonbonded_forces,
+        include_bond_forces=cgmodel.include_bond_forces,
+        include_bond_angle_forces=cgmodel.include_bond_angle_forces,
+        include_torsion_forces=cgmodel.include_torsion_forces,
+        constrain_bonds=cgmodel.constrain_bonds,
+        sequence=cgmodel.sequence,
+        positions=cgmodel.positions,
+        monomer_types=cgmodel.monomer_types,
+    )
+    
+    native_structure_file=f"{structures_path}/medoid_0.dcd"
+
+    native_traj = md.load(native_structure_file,top=md.Topology.from_openmm(cgmodel.topology))    
+    
+    positions = native_traj.xyz[0] * unit.nanometer
+    
+    output_directory = tmpdir.mkdir("output")
+    
+    # Minimize energy of native structure
+    positions, PE_start, PE_end, simulation = minimize_structure(
+        cgmodel_new,
+        positions,
+        output_file=f"{output_directory}/medoid_min.dcd",
+    )
+    
+    # These should be equal to ~4 decimal places (1 Joule/mol)
+    PE_start_kappa_off = -382.19839163767057
+    PE_end_kappa_off = -500.99943208890255
+    
+    PE_start_kappa_on = PE_start.value_in_unit(unit.kilojoule_per_mole)
+    PE_end_kappa_on = PE_end.value_in_unit(unit.kilojoule_per_mole)
+    
+    assert_almost_equal(PE_start_kappa_on,PE_start_kappa_off,decimal=4)
+    assert_almost_equal(PE_end_kappa_on,PE_end_kappa_off,decimal=4)
+    
+    
+def test_set_binary_interaction(tmpdir):
+    """Test for adding binary interaction parameters"""
+
+    # Load in cgmodel
+    cgmodel = pickle.load(open(f"{data_path}/stored_cgmodel.pkl", "rb" ))    
+
+    binary_interaction_parameters = {
+        "bb_sc_binary_interaction": 0.25}
+    
+    cgmodel_new = CGModel(
+        particle_type_list=cgmodel.particle_type_list,
+        bond_lengths=cgmodel.bond_lengths,
+        bond_force_constants=cgmodel.bond_force_constants,
+        bond_angle_force_constants=cgmodel.bond_angle_force_constants,
+        torsion_force_constants=cgmodel.torsion_force_constants,
+        equil_bond_angles=cgmodel.equil_bond_angles,
+        torsion_periodicities=cgmodel.torsion_periodicities,
+        torsion_phase_angles=cgmodel.torsion_phase_angles,
+        binary_interaction_parameters=binary_interaction_parameters,
+        include_nonbonded_forces=cgmodel.include_nonbonded_forces,
+        include_bond_forces=cgmodel.include_bond_forces,
+        include_bond_angle_forces=cgmodel.include_bond_angle_forces,
+        include_torsion_forces=cgmodel.include_torsion_forces,
+        constrain_bonds=cgmodel.constrain_bonds,
+        sequence=cgmodel.sequence,
+        positions=cgmodel.positions,
+        monomer_types=cgmodel.monomer_types,
+    )   
+    
+    
+def test_set_Mie_exponents_trivial(tmpdir):
+    """
+    Regression test for using Mie potential customNonbondedForce, with 12-6 exponents,
+    no binary interaction parameter
+    """
+
+    # Load in cgmodel
+    cgmodel = pickle.load(open(f"{data_path}/stored_cgmodel.pkl", "rb" ))    
+    
+    # Set Mie exponents:
+    n = 12
+    m = 6
+    
+    cgmodel_new = CGModel(
+        particle_type_list=cgmodel.particle_type_list,
+        bond_lengths=cgmodel.bond_lengths,
+        bond_force_constants=cgmodel.bond_force_constants,
+        bond_angle_force_constants=cgmodel.bond_angle_force_constants,
+        torsion_force_constants=cgmodel.torsion_force_constants,
+        equil_bond_angles=cgmodel.equil_bond_angles,
+        torsion_periodicities=cgmodel.torsion_periodicities,
+        torsion_phase_angles=cgmodel.torsion_phase_angles,
+        nonbond_repulsive_exp=n,
+        nonbond_attractive_exp=m,
+        include_nonbonded_forces=cgmodel.include_nonbonded_forces,
+        include_bond_forces=cgmodel.include_bond_forces,
+        include_bond_angle_forces=cgmodel.include_bond_angle_forces,
+        include_torsion_forces=cgmodel.include_torsion_forces,
+        constrain_bonds=cgmodel.constrain_bonds,
+        sequence=cgmodel.sequence,
+        positions=cgmodel.positions,
+        monomer_types=cgmodel.monomer_types,
+    )
+    
+    native_structure_file=f"{structures_path}/medoid_0.dcd"
+
+    native_traj = md.load(native_structure_file,top=md.Topology.from_openmm(cgmodel.topology))    
+    
+    positions = native_traj.xyz[0] * unit.nanometer
+    
+    output_directory = tmpdir.mkdir("output")
+    
+    # Minimize energy of native structure
+    positions, PE_start, PE_end, simulation = minimize_structure(
+        cgmodel_new,
+        positions,
+        output_file=f"{output_directory}/medoid_min.dcd",
+    )
+    
+    # Check that the energy matches the standard LJ 12-6 potential:
+    # These should be equal to ~4 decimal places (1 Joule/mol)
+    PE_start_LJ_12_6 = -382.19839163767057
+    PE_end_LJ_12_6 = -500.99943208890255
+    
+    PE_start_Mie_12_6 = PE_start.value_in_unit(unit.kilojoule_per_mole)
+    PE_end_Mie_12_6 = PE_end.value_in_unit(unit.kilojoule_per_mole)
+    
+    assert_almost_equal(PE_start_LJ_12_6,PE_start_Mie_12_6,decimal=4)
+    assert_almost_equal(PE_end_LJ_12_6,PE_end_Mie_12_6,decimal=4) 
+
+
+def test_set_Mie_exponents(tmpdir):
+    """
+    Test for using Mie potential customNonbondedForce,
+    no binary interaction parameter
+    """
+
+    # Load in cgmodel
+    cgmodel = pickle.load(open(f"{data_path}/stored_cgmodel.pkl", "rb" ))    
+    
+    # Set Mie exponents:
+    n = 14.0
+    m = 7.0
+    
+    cgmodel_new = CGModel(
+        particle_type_list=cgmodel.particle_type_list,
+        bond_lengths=cgmodel.bond_lengths,
+        bond_force_constants=cgmodel.bond_force_constants,
+        bond_angle_force_constants=cgmodel.bond_angle_force_constants,
+        torsion_force_constants=cgmodel.torsion_force_constants,
+        equil_bond_angles=cgmodel.equil_bond_angles,
+        torsion_periodicities=cgmodel.torsion_periodicities,
+        torsion_phase_angles=cgmodel.torsion_phase_angles,
+        nonbond_repulsive_exp=n,
+        nonbond_attractive_exp=m,
+        include_nonbonded_forces=cgmodel.include_nonbonded_forces,
+        include_bond_forces=cgmodel.include_bond_forces,
+        include_bond_angle_forces=cgmodel.include_bond_angle_forces,
+        include_torsion_forces=cgmodel.include_torsion_forces,
+        constrain_bonds=cgmodel.constrain_bonds,
+        sequence=cgmodel.sequence,
+        positions=cgmodel.positions,
+        monomer_types=cgmodel.monomer_types,
+    )
+
+
+def test_set_Mie_exponents_and_binary_parameter_trivial(tmpdir):
+    """
+    Regression test for using Mie potential customNonbondedForce, with 12-6 exponents,
+    with a binary interaction parameter for epsilon mixing rules.
+    Currently the nontrivial case is not yet implemented.
+    """
+
+    # Load in cgmodel
+    cgmodel = pickle.load(open(f"{data_path}/stored_cgmodel.pkl", "rb" ))    
+    
+    # Set Mie exponents:
+    n = 12
+    m = 6
+    
+    # Set binary interaction parameter:
+    binary_interaction_parameters = {
+        "bb_sc_binary_interaction": 0.0}    
+    
+    cgmodel_new = CGModel(
+        particle_type_list=cgmodel.particle_type_list,
+        bond_lengths=cgmodel.bond_lengths,
+        bond_force_constants=cgmodel.bond_force_constants,
+        bond_angle_force_constants=cgmodel.bond_angle_force_constants,
+        torsion_force_constants=cgmodel.torsion_force_constants,
+        equil_bond_angles=cgmodel.equil_bond_angles,
+        torsion_periodicities=cgmodel.torsion_periodicities,
+        torsion_phase_angles=cgmodel.torsion_phase_angles,
+        nonbond_repulsive_exp=n,
+        nonbond_attractive_exp=m,
+        binary_interaction_parameters=binary_interaction_parameters,
+        include_nonbonded_forces=cgmodel.include_nonbonded_forces,
+        include_bond_forces=cgmodel.include_bond_forces,
+        include_bond_angle_forces=cgmodel.include_bond_angle_forces,
+        include_torsion_forces=cgmodel.include_torsion_forces,
+        constrain_bonds=cgmodel.constrain_bonds,
+        sequence=cgmodel.sequence,
+        positions=cgmodel.positions,
+        monomer_types=cgmodel.monomer_types,
+    )
+    
+    native_structure_file=f"{structures_path}/medoid_0.dcd"
+
+    native_traj = md.load(native_structure_file,top=md.Topology.from_openmm(cgmodel.topology))    
+    
+    positions = native_traj.xyz[0] * unit.nanometer
+    
+    output_directory = tmpdir.mkdir("output")
+    
+    # Minimize energy:
+    positions, PE_start, PE_end, simulation = minimize_structure(
+        cgmodel_new,
+        positions,
+        output_file=f"{output_directory}/medoid_min.dcd",
+    )
+    
+    # Check that the energy matches the standard LJ 12-6 potential:
+    # These should be equal to ~4 decimal places (1 Joule/mol)
+    PE_start_LJ_12_6 = -382.19839163767057
+    PE_end_LJ_12_6 = -500.99943208890255
+    
+    PE_start_Mie_12_6 = PE_start.value_in_unit(unit.kilojoule_per_mole)
+    PE_end_Mie_12_6 = PE_end.value_in_unit(unit.kilojoule_per_mole)
+    
+    assert_almost_equal(PE_start_LJ_12_6,PE_start_Mie_12_6,decimal=4)
+    assert_almost_equal(PE_end_LJ_12_6,PE_end_Mie_12_6,decimal=4)     
+
